refactor: Waiting을 Reservation Controller에서 분리
