+++ src/test/java/roomescape/acceptance/ReservationAcceptanceTest.java
@@ -34,28 +34,6 @@ class ReservationAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("예약 대기를 생성한다.")
-    void createWaitingTest() {
-        long themeId = fixture.createTheme(new ThemeRequest("name", "desc", "url")).id();
-        long timeId = fixture.createReservationTime(10, 0).id();
-        fixture.registerMember(new MemberRegisterRequest("name", "email@mail.com", "12341234"));
-        String token = fixture.loginAndGetToken("email@mail.com", "12341234");
-        fixture.createReservation(token, new ReservationRequest(LocalDate.of(2024, 12, 25), timeId, themeId));
-
-        fixture.registerMember(new MemberRegisterRequest("test", "test@mail.com", "12341234"));
-        String waitingToken = fixture.loginAndGetToken("test@mail.com", "12341234");
-
-        ReservationRequest waitingRequest = new ReservationRequest(LocalDate.of(2024, 12, 25), timeId, themeId);
-        RestAssured.given().log().all()
-                .cookie("token", waitingToken)
-                .contentType(ContentType.JSON)
-                .body(waitingRequest)
-                .when().post("/reservations/waiting")
-                .then().log().all()
-                .statusCode(201);
-    }
-
-    @Test
     @DisplayName("관리자가 예약을 삭제한다.")
     void deleteReservation() {
         long themeId = fixture.createTheme(new ThemeRequest("name", "desc", "url")).id();
@@ -70,29 +48,6 @@ class ReservationAcceptanceTest extends AcceptanceTest {
         RestAssured.given().log().all()
                 .cookie("token", token)
                 .when().delete("/reservations/{id}", response.id())
-                .then().log().all()
-                .statusCode(204);
-    }
-
-    @Test
-    @DisplayName("사용자가 예약 대기를 삭제한다.")
-    void deleteWaiting() {
-        long themeId = fixture.createTheme(new ThemeRequest("name", "desc", "url")).id();
-        long timeId = fixture.createReservationTime(10, 0).id();
-
-        fixture.registerMember(new MemberRegisterRequest("name", "email@mail.com", "12341234"));
-        String reservationToken = fixture.loginAndGetToken("email@mail.com", "12341234");
-        fixture.createReservation(reservationToken,
-                new ReservationRequest(LocalDate.of(2024, 12, 25), timeId, themeId));
-
-        fixture.registerMember(new MemberRegisterRequest("test", "test@mail.com", "12341234"));
-        String waitingToken = fixture.loginAndGetToken("test@mail.com", "12341234");
-        ReservationResponse response = fixture.createWaiting(waitingToken,
-                new ReservationRequest(LocalDate.of(2024, 12, 25), timeId, themeId));
-
-        RestAssured.given().log().all()
-                .cookie("token", waitingToken)
-                .when().delete("/reservations/waiting/{id}", response.id())
                 .then().log().all()
                 .statusCode(204);
     }
@@ -115,15 +70,5 @@ class ReservationAcceptanceTest extends AcceptanceTest {
                 .as(ReservationStatusResponse[].class);
 
         assertThat(responses).hasSize(1);
-    }
-
-    @Test
-    @DisplayName("관리자가 예약 목록을 조회한다.")
-    void adminReadWaiting() {
-        RestAssured.given().log().all()
-                .cookie("token", fixture.getAdminToken())
-                .when().get("admin/reservations/waiting")
-                .then().log().all()
-                .statusCode(200);
     }
 }
